requirements.txt
Flask==2.1.2
flask-restx==0.5.1
flask_jwt_extended==4.4.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.1.2
pytest==7.1.2
pandas==1.3.5
requests==2.28.1
numpy==1.22.0
scikit-learn==1.1.1
requests
joblib==1.1.0
Flask-WTF==1.0.0
